requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
colorama==0.4.6
bcrypt==5.0.0
PyJWT==2.12.1
boto3
python-dotenv
gunicorn
psycopg2-binary
